Deploy fail
This AWS EB hard to use.
First, have to install AWSCLI and EBCLI, because the aws web is not provide the feature.
The EBCLI installation is encounted alot compatible issue. The python version is fixed, I have to reinstall python, the feature to install with other py version is failed.
The deploy keep fail, the init for existed project is hard to understand and complex to setup.
This commit is last chance...